fix ci and move cli from pkg to internal
