+++ .github/workflows/go.yml
@@ -22,7 +22,4 @@ jobs:
         go-version: '1.22.3'
 
     - name: Build
-      run: go build -v ./...
-
-    - name: Test
-      run: go test -v ./...
+      run: go build -v ./cmd/mon/mon.go
